修改阿克曼，使用tan表并弃用浮点数；修改pid控制器定义，全部弃用float使用int16类型；弃用adc、imu、舵机中多余代码；调整车轮对浮点类型的使用，改用int16

+++ project/code/dev/dev_wheel.h
@@ -23,17 +23,17 @@ typedef enum
 // 左右轮PID控制器
 extern pid_control_t wheel_pid_left;
 extern pid_control_t wheel_pid_right;
-extern float car_wheel_target_speed;
-extern float ref_left_target;
-extern float ref_right_target;
+extern int16 car_wheel_target_speed;
+extern int16 ref_left_target;
+extern int16 ref_right_target;
 
-// 函数 
+// 函数
 void car_wheel_set_speed(car_wheel_index_enum wheel, int8 speed_percent); // 单独设置速度
 void car_wheel_set_dual(int8 right_wheel_speed_percent, int8 left_wheel_speed_percent); // 设置占空比
 void car_wheel_stop_all(void);  // 全部停止
 void car_wheel_init(void); // 初始化
 void car_wheel_pid_init(void); // pid结构体初始化
-void car_wheel_set_target(float speed); // 设置整车目标速度，由内部自动换算左右轮目标
+void car_wheel_set_target(int16 speed); // 设置整车目标速度，由内部自动换算左右轮目标
 void car_wheel_control_reset(void); // 清空闭环状态并停轮
 void car_wheel_hold(void); // 保持后轮待转，不清 Start 速度目标
 void car_wheel_update(void); // 自动调用pid进行车轮更新函数
